reorganize startpos switcher
